Add files via upload
A digital clock means it shows time in number format. JavaScript receives real-time out of your tool and presentations it at the website. And this post is about how JavaScript gets the time and displays it using some basic methods. Basically, this is a real-time digital clock, which means you can see time in numbers with hours, minutes, and second. And the clock usually works, so you don't need to refresh the website to see the updated time.


Let's start making these amazing responsive navbar with hamburger menu Using HTML, CSS and JavaScript step by step.